feat: auth admin user with token

+++ routes/api.php
@@ -7,8 +7,8 @@ Route::post("/users", [UsersController::class, 'storeUser']);
 Route::post("/users/login", [UsersController::class, 'login']);
 
 Route::middleware(['auth.custom'])->group(function () {
-    Route::get("/users", [UsersController::class, 'getUsers']);
-    Route::get("/users/{id}", [UsersController::class, 'retrieveUser']);
+    Route::get("/users", [UsersController::class, 'getUsers'])->middleware('admin');
+    Route::get("/users/{id}", [UsersController::class, 'retrieveUser'])->middleware('admin');
     Route::patch("/users/{id}", [UsersController::class, 'updateUser']);
-    Route::delete("/users/{id}", [UsersController::class, 'destroyUser']);
+    Route::delete("/users/{id}", [UsersController::class, 'destroyUser'])->middleware('admin');
 });
